#2 Minimum_Swaps_2 passing solution
Original solution worked in O(n^2) time, while this one should work in O(n). 
We organise the values into multiple cycles of swaps to be done. Number of needed swaps is a sum of the cycle sizes with each decreased by one (because it is a cycle after all).